Modal: move to shared folder
